requirements.txt
Flask>=2.3,<3.0
pandas>=2.0,<3.0
numpy>=1.23,<2.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
duckdb>=0.8.0
adlfs>=2.15.0
pyarrow>=14.0.0